Extract triangulum execution into module
