+++ gem/lib/triangulum.rb
@@ -9,4 +9,5 @@ module Triangulum
 end
 
 require_relative 'triangulum/version'
+require_relative 'triangulum/executor'
 require_relative 'triangulum/validation'
